requirement.txt
requests
python-dotenv
beautifulsoup4
cohere
